Check for znapzendzetup before trying to run it
If znapzend isn't installed, the file systems table shows "Error" in
every row for the Replication column as it tries znapzendzetup for each.

This updates it to first check it's available, and showing "Unavailable"
when it's not to reduce confusion. It will only be "Error" if the
`znapzendzetup` command actually runs and fails for some reason.